Split up ProfileWriter
